Moved music to dict to music dataclass. This made more sense for organization
